+++ src/include/network/network_defs.h
@@ -1,17 +1,5 @@
 #pragma once
 
-#include <unistd.h>
-
-#include <bitset>
-#include <climits>
-#include <cstdint>
-#include <functional>
-#include <limits>
-#include <map>
-#include <memory>
-#include <string>
-#include <unordered_map>
-#include <unordered_set>
 #include <vector>
 
 #include "common/macros.h"
